fix: Change recipe image type to string instead of Image
Change in recipe model image type to string (image url) instead of Image (FileType/Binary). This is decided as currently hosting on vercel does not allow for a seamless data cloud storage integration. We prefer adding the image on a cloud storage (or google image) and adding its link to the recipe